Partially geocoded 911 data - output of mapping script
In order to use the Mapping_911_Jan2019 R script it requires a personalized and project-specific Google Maps API key. There are monetary costs associated with running the script so I am including my results of running the script. 
The script hit an error after about 5000 rows and then it displays the same value after that point. It is still useful to have the rows that worked. With time I may comeback to debug it.